moved the fetch profile up one level and fixed the check for venueManager
