+++ src/stores/useAuthStore.js
@@ -10,14 +10,14 @@ const useAuthStore = create(
       isLoggedIn: false,
       accessToken: null,
       userName: null,
-      venueManager: false,
+      // venueManager: false,
       // user: null,
 
       setIsLoggedIn: () => set({ isLoggedIn: true }),
       setIsLoggedOut: () => set({ isLoggedIn: false }),
       setUserName: (name) => set({ userName: name }),
       setAccessToken: (token) => set({ accessToken: token }),
-      setVenueManager: (bool) => set({ venueManager: bool }),
+      // setVenueManager: (bool) => set({ venueManager: bool }),
       // setUser: (user) => set({ user: user }),
       logOut: () => set({ isLoggedIn: false, accessToken: null, userName: null, user: null }),
     }),
